+++ src/components/index.js
@@ -3,3 +3,4 @@ import 'react-virtualized/styles.css';
 export { VirtualizedTable } from './VirtualizedTable';
 export { MultiGridTable } from './MultiGridTable';
 export { ReactSortableTree } from './ReactSortableTree';
+export { VirtualizedSelect } from './VirtualizedSelect';