+++ src/utils/index.js
@@ -1,1 +1,2 @@
 export { makeDelaySearch } from './makeDelaySearch';
+export { search } from './search';